Add book spine segmentation module to src
